refactor: Move BackgroundConnection to package root to be sharable
